Remove code that checks for luamod submodule
There was some code in the external library CMake glue code that would check if the luamod submodule exists, and if not invoke git to init all the submodules automatically.

Since we removed the luamod dependency (and hence all git submodules) we can get rid of this code. It also stops CMake trying to clone submodules that don't exist on configure.