Update target C++ standard to C++17
As previously discussed we'll now target C++17 instead of just C++11, so
this updates the CMake to reflect that change.

This also gets rid of manually adding --std=xxxxx for Unix targets. This
should not be needed as setting CMAKE_CXX_STANDARD should do that for
us. By manually adding --std=xxxxx the compiler invocation would
probably end up with multiple arguments specifying the language version.